requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
aiosqlite==0.20.0
aioschedule==0.5.2
python-dotenv==1.0.1
